move elementCenterPoint from utils to bounds.ts
