+++ packages/element/src/binding.ts
@@ -6,7 +6,6 @@ import {
   invariant,
   isDevEnv,
   isTestEnv,
-  elementCenterPoint,
 } from "@excalidraw/common";
 
 import {
@@ -61,7 +60,7 @@ import {
   isTextElement,
 } from "./typeChecks";
 
-import { aabbForElement } from "./bounds";
+import { aabbForElement, elementCenterPoint } from "./bounds";
 import { updateElbowArrowPoints } from "./elbowArrow";
 
 import type { Scene } from "./Scene";
