+++ packages/excalidraw/renderer/helpers.ts
@@ -1,8 +1,8 @@
-import { elementCenterPoint, THEME, THEME_FILTER } from "@excalidraw/common";
+import { THEME, THEME_FILTER } from "@excalidraw/common";
 
 import { FIXED_BINDING_DISTANCE } from "@excalidraw/element";
 import { getDiamondPoints } from "@excalidraw/element";
-import { getCornerRadius } from "@excalidraw/element";
+import { elementCenterPoint, getCornerRadius } from "@excalidraw/element";
 
 import {
   curve,
